+++ src-tauri/src/models.rs
@@ -272,6 +272,7 @@ pub(crate) struct RemoteWorkspaceConfig {
     pub(crate) remote_zomboid_data_owner: String,
     pub(crate) remote_client_ram: String,
     pub(crate) remote_server_ram: String,
+    pub(crate) remote_server_version: String,
     pub(crate) remote_setup_completed_step: u8,
     pub(crate) remote_mod_locations: Vec<String>,
 }